requirements.txt
streamlit
requests
python-dotenv
fastapi
uvicorn[standard]
langchain
langchain-google-genai
pypdf
faiss-cpu
pydantic-settings
python-dotenv
python-multipart
langchain_community
pymupdf
fitz
frontend
tools
protobuf==3.20.3